hot fix: 1. discount was amended to double 2. product validator was amended from else if to if 3. product validator formating was applied

+++ src/com/javaguru/shoppinglist/CreateProductAction.java
@@ -25,7 +25,7 @@ public class CreateProductAction implements Action {
         System.out.println("Enter product price: ");
         String price = scanner.nextLine();
         System.out.println("Enter product discount: ");
-        Long discount = scanner.nextLong();
+        Double discount = scanner.nextDouble();
 
 
         Product product = new Product();
